requirements.txt
websocket_client
numpy
TA-Lib
python-binance
pyTelegramBotAPI
fire